refactor(react-e-commerce/server): move category helpers into the category file
